% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGATnb.R
\name{aggregateGATnb}
\alias{aggregateGATnb}
\title{Aggregate GAT neighbors}
\usage{
aggregateGATnb(nb, ids, remove.self = TRUE)
}
\arguments{
\item{nb}{Neighbors object.}

\item{ids}{Vector of IDs corresponding to the neighbor object's
"region.id" attribute.}

\item{remove.self}{Boolean denoting whether to remove original areas.}
}
\description{
This function is a clone of spdep::aggregate.nb with one
minor modification - it does not remove areas without neighbors.
}
