% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGATcompactness.R
\name{calculateGATcompactness}
\alias{calculateGATcompactness}
\title{Calculate GAT Compactness}
\usage{
calculateGATcompactness(shp)
}
\arguments{
\item{shp}{A spatial polygons data frame.}
}
\description{
This function calculates the compactness ratios for the aggregated areas.

It checks whether the projection of a spatial polygons data
frame is in lat/long format. If the check passes, the function assigns
an alternate projection. The spatial polygons data frame is then used
to calculate compactness ratio. The function returns a vector of ratios.
}
\examples{
calculateGATcompactness(
  shp = hftown
)

}
