% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGATexclusions.R
\name{calculateGATexclusions}
\alias{calculateGATexclusions}
\title{Calculate GAT Exclusions}
\usage{
calculateGATexclusions(d, var, math, val)
}
\arguments{
\item{d}{The dataset containing the flag and exclusion variables.}

\item{var}{The variable to use when calculating exclusions.}

\item{math}{The condition to use when calculating exclusions. Recognized
conditions include {"equals", "less than", "greater than"}.}

\item{val}{The value to compare to \code{var} based on the condition selected.}
}
\description{
This function cleans up the code to flag areas based on exclusion
criteria so that they will not be used in the merge.
}
\examples{

calculateGATexclusions(
  d = hftown@data,
  var = "TOTAL_POP",
  math = "less than",
  val = 500
)

}
