% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGATflag.R
\name{calculateGATflag}
\alias{calculateGATflag}
\title{Calculate GAT Flag}
\usage{
calculateGATflag(exclist, d)
}
\arguments{
\item{exclist}{The vector of settings to use, including variable names
{"var1", "var2", "var3"}, the criteria {"math1", "math2",
"math3"}, and the numeric values {"val1", "val2", "val3"}.}

\item{d}{The dataset containing the flag and exclusion variables.}
}
\description{
This function cleans up the code to update the flag variable based on
exclusion criteria.

I will need to add an example later.
}
