% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGATshapefilesize.R
\name{checkGATshapefilesize}
\alias{checkGATshapefilesize}
\title{Check GAT Shapefile Size}
\usage{
checkGATshapefilesize(userfile)
}
\arguments{
\item{userfile}{A filepath to a shapfile, including filename, without the
extension.}
}
\description{
This function is used to determine whether the shapefile is too large for
GAT and the user's computer to handle it. This function makes a very
conservative estimate that may be refined later.

This function returns a number, 0 or 100. If it returns 0, the check passed
and you can use the file. If it returns 100, the check failed and you will
need to reduce the size or choose a new shapefile.
}
\details{
Note from Gwen: "I am not sure of the relationship between the size of the
R objects produced by file import and the file size. We are assumming 1:1
here, but this may not be correct. Size appears to increase at least 10x
going from disk to memory."
}
\examples{

\dontrun{
# locate the example shapefile in the package
myfile <- paste0(find.package("gatpkg"), "/extdata/hftown")

# check the memory size of the shapefile
# user only receives a message if there is a problem
checkGATshapefilesize(
  myfile
)
}

}
