% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGATvariabletypes.R
\name{checkGATvariabletypes}
\alias{checkGATvariabletypes}
\title{Check GAT Variable Types}
\usage{
checkGATvariabletypes(mapdata, type = "number")
}
\arguments{
\item{mapdata}{A data frame, intended to be read from a shapefile DBF.}

\item{type}{Variable type. Options include "number", which returns a vector
of names for all numeric variables, and "character", which
returns a vector of names for all character variables.}
}
\description{
This function is used to create a list of either numeric or character
variables in the data frame.
}
\examples{
# identify numeric variables
my_numericvars <-
  checkGATvariabletypes(
    mapdata = hftown,
    type = "number"
  )

}
