% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGATcrosswalks.R
\name{combineGATcrosswalks}
\alias{combineGATcrosswalks}
\title{Combine GAT Crosswalks}
\usage{
combineGATcrosswalks(path, file1, file2, idvar)
}
\arguments{
\item{path}{The filepath to the files created by GAT. This function
assumes the same filepath was used in both runs.}

\item{file1}{The crosswalk (*in) file created in the first run of GAT.
For best results, avoid filenames containing a period. It is
not necessary to add the file extension.}

\item{file2}{The crosswalk (*in) file created in the second run of GAT.
For best results, avoid filenames containing a period. It is
not necessary to add the file extension.}

\item{idvar}{The ID variable used when running GAT. This function assumes
the same ID variable was used in both runs.}
}
\description{
This function provides complete crosswalk and final shapefiles in cases
when the user requires GAT to be run twice (for example, a first run that
enforces the boundary variable and a second run that does not).
}
\details{
Two shapefiles are created and saved in the path provided:
\itemize{
\item a revised original crosswalk
\itemize{
\item name: file1 ending in "_combined"
\item "GATid" renamed to "tempid"
\item new "GATid" created with the IDs from the second run
}
\item a revised final aggregation
\itemize{
\item name: file2 without "in" ending in "_combined"
\item "GATnumIDs" removed
\item new "GATnumIDs" created with the number of areas in the original
shapefile included in each final area
}
}

In a future iteration, I may also include maps and a log that
compare the original and final shapefiles.
}
