% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertlatlong2UTM.R
\name{convertlatlong2UTM}
\alias{convertlatlong2UTM}
\title{Latitude/Longitude to UTM Coordinates}
\usage{
convertlatlong2UTM(area, units = "m")
}
\arguments{
\item{area}{A SpatialPolygonsDataFrame.}

\item{units}{The measurement unit. Default is meters.}
}
\description{
This function converts scalar to planar coordinates
}
\examples{
convertlatlong2UTM(area = hftown, units = "m")


}
