% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGATregion.R
\name{createGATregion}
\alias{createGATregion}
\title{Create New GAT Region}
\usage{
createGATregion(
  mydata,
  newreg,
  myidvar,
  nrid,
  area = NULL,
  pop = NULL,
  IDlist = NULL,
  pwrepeat = FALSE,
  popwt = FALSE
)
}
\arguments{
\item{mydata}{A data frame from the original shapefile.}

\item{newreg}{The rows in the data frame (polygons in the shapefile later)
that will be joined into a single observation (polygon).}

\item{myidvar}{A variable of unique string values to identify the
observations.}

\item{nrid}{A string that will be the ID for the merged observation
(polygon).}

\item{area}{The area shapefile (if pwrepeat=TRUE or GATx/GATy missing).}

\item{pop}{The base population shapefile (if popwt=TRUE).}

\item{IDlist}{The list to match GAT and old IDs (if popwt=TRUE).}

\item{pwrepeat}{A boolean denoting whether population weighting (if used)
should be recalculated each time two areas are merged
(TRUE) or if area centroids should be weighted with area
populations (FALSE). If population weighting is not used,
this option is ignored.}

\item{popwt}{A boolean that denotes whether population weighting is
being used.}
}
\description{
This function combines multiple observations by summing the numeric
variables, assigning one value for the character variables, and replacing
the ID value with a unique ID. It returns the resulting single observation.
}
\details{
This function can handle more than two observations at a time, but has only
been tested with two observations.
}
\examples{
# create dataset with polygons to merge
ids <- c("e", "e", "d", "d", "f", "f", "f", "e", "e", "d", "e",
         "d", "b", "c", "c", "a", "b", "c", "a", "b", "c")
my_merged_obs <-
  createGATregion(
    mydata = hftown@data,
    newreg = hftown@data[ids=="a",],
    myidvar = "ID",
    nrid = "a",
    area = hftown,
    pop = hfpop,
    IDlist = ids,
    pwrepeat = TRUE,
    popwt = TRUE
  )

}
