% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineGATmapclasses.R
\name{defineGATmapclasses}
\alias{defineGATmapclasses}
\title{Define Classes for Map Legends}
\usage{
defineGATmapclasses(areaold, areanew, aggvar, breaks = 7)
}
\arguments{
\item{areaold}{A spatial polygons data frame.}

\item{areanew}{A second spatial polygons data frame that ahould have the
same outer boundary as the first one.}

\item{aggvar}{The variable to be mapped.}

\item{breaks}{The number of desired categories in the maps.}
}
\description{
This function prepares the classes for the original and aggregated maps
so that the same scale can be used for both.
}
\examples{
# prepare the classes for the choropleth maps
defineGATmapclasses(
  areaold = hftown,
  areanew = hfagg610k,
  aggvar = "TOTAL_POP",
  breaks = 4
)

}
