% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatpkg.R
\docType{package}
\name{gatpkg-package}
\alias{gatpkg}
\alias{gatpkg-package}
\title{gatpkg: The NYS Geographic Aggregation Tool (GAT)}
\description{
GAT provides a way of easily aggregating sptial data using user-defined
parameters without requiring much knowledge of or experience using R.
}
\details{
Its goals:
\itemize{
\item simplify geographic aggregation
\item be in a format accessible to anyone
\item not require prior experience with R
\item be customizable for advanced R users
}

To run the default version of GAT, follow the GAT user's manual in the
vignettes: \verb{praceholder - check link}.

For further information about GAT ... \code{placeholder}
}
\author{
\strong{Maintainer}: Abigail Stamm \email{abigail.stamm@health.ny.gov} (author of R package)

Authors:
\itemize{
  \item Gwen Babcock \email{gdb02@health.state.ny.us} (author of SAS and R scripts) [translator]
}

Other contributors:
\itemize{
  \item Thomas Talbot \email{tot01@health.state.ny.us} (created the original concept of GAT) [contributor]
  \item Arjita Rai \email{arjita.rai@health.ny.gov} (alpha tester, provided functionality suggestions) [contributor]
  \item Douglas Done \email{douglas.done@health.ny.gov} (provided functionality suggestions) [contributor]
  \item Neil Muscatiello \email{neil.muscatiello@health.ny.gov} (alpha tester) [contributor]
  \item Marine Nalbandyan \email{marine.nalbandyan@health.ny.gov} (alpha tester) [contributor]
  \item Catherine Wargo \email{catherine.wargo@health.ny.gov} (alpha tester) [contributor]
  \item Maxime Fokoua Dongmo \email{maxime.fokoua@health.ny.gov} (alpha tester) [contributor]
  \item NYSDOH Environmental Public Health Tracking \email{epht@health.ny.gov} [copyright holder]
}

}
