% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatpkg.R
\docType{package}
\name{gatpkg}
\alias{gatpkg}
\alias{_PACKAGE}
\alias{GAT}
\title{The NYS Geographic Aggregation Tool}
\description{
A method for easily aggregating spatial data using user-defined
parameters with minimal knowledge of or experience using R.
}
\details{
GAT's goals:
\itemize{
\item simplify geographic aggregation
\item be in a format accessible to anyone
\item not require prior experience with R
\item be customizable for advanced R users
}

To run the default version of GAT, follow the GAT tutorial
\href{../doc/gat_tutorial.html}{GAT tutorial} in the vignettes.

For further information about GAT ... \code{placeholder}

GAT was funded by the CDC NEPHT grant #15-0910-03. Its content is solely
the responsibility of the authors and does not necessarily represent the
official views of the CDC.
}
\author{
\bold{Authors} \cr
Abigail Stamm, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Gwen Babcock, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr

\bold{Alpha-testers} \cr
Neil Muscatiello, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Arjita Rai, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Catherine Wargo, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Christophe Maxime Fokoua Dongmo, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Marine Nalbandyan, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr

\bold{Other contributors} \cr
Douglas Done, Bureau of Environmental and Occupational Epidemiology,
New York State Department of Health \cr
Thomas Talbot, Bureau of Environmental and Occupational Epidemiology,
New York State Department of Health (retired) \cr
}
