% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfagg5k.R
\docType{data}
\name{hfagg5k}
\alias{hfagg5k}
\title{Hamilton & Fulton counties aggregated spatial polygon data frame}
\format{
A SpatialPolygonsDataFrame object that includes the following
11 variables:

\describe{
\item{\code{ID}}{Census county and town FIPS designation}
\item{\code{TOWN}}{Town name}
\item{\code{COUNTY}}{Census county FIPS designation}
\item{\code{AREALAND}}{Land area}
\item{\code{AREAWATR}}{Water area}
\item{\code{TOTAL_POP}}{Census population for 2010}
\item{\code{MY_FLAG}}{Arbitrary flag variable to illustrate exclusions}
\item{\code{GATflag}}{Placeholder for excluded areas, if any}
\item{\code{GATx}}{Recalculated longitude}
\item{\code{GATy}}{Recalculated latitude}
\item{\code{GATnumIDs}}{number of towns aggregated to the area}
}
}
\usage{
hfagg5k
}
\description{
This object was aggregated from a US census shapefile of towns in Hamilton
and Fulton Counties in New York State. The settings used were:
\itemize{
\item aggregate total population to at least 5,000
\item maintain county borders
\item aggregate by similar ratio of land to water
\item do not exclude any areas
}
}
\keyword{datasets}
