% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfagg.R
\docType{data}
\name{hfagg615k}
\alias{hfagg615k}
\title{Hamilton & Fulton counties aggregated to 15k population}
\format{
Simple features object with the following structure.

\describe{
\item{\code{ID}}{Census county and town FIPS designation}
\item{\code{TOWN}}{Town name}
\item{\code{COUNTY}}{Census county FIPS designation}
\item{\code{TOTAL_POP}}{Census population for 2010}
\item{\code{GATflag}}{Flag values for excluded and problematic areas, if any}
\item{\code{GATx}}{Recalculated longitude}
\item{\code{GATy}}{Recalculated latitude}
\item{\code{GATnumIDs}}{number of towns aggregated to the area}
}
}
\usage{
hfagg615k
}
\description{
This object was created by aggregating a US census shapefile of towns
in Hamilton and Fulton Counties in New York State. The settings used were:
\itemize{
\item aggregate hfagg610k total population to at least 6,000 and at most
15,000
\item request county borders, but do not require them
\item aggregate by closest geographic centroid
\item do not exclude any areas
}
}
\keyword{datasets}
