% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfagg.R
\docType{data}
\name{hfcw610k}
\alias{hfcw610k}
\title{Hamilton & Fulton counties aggregated simple feature}
\format{
Simple features object with the following structure.

\describe{
\item{\code{ID}}{Census county and town FIPS designation}
\item{\code{TOWN}}{Town name}
\item{\code{COUNTY}}{Census county FIPS designation}
\item{\code{TOTAL_POP}}{Census population for 2010}
\item{\code{GATflag}}{Flag values for excluded areas, if any}
\item{\code{GATid}}{GAT-generated ID of assigned aggregated area}
}
}
\usage{
hfcw610k
}
\description{
This object was created by aggregating a US census shapefile of towns
in Hamilton and Fulton Counties in New York State. The settings used were:\preformatted{* aggregate hftown total population to at least 6,000 and at most 10,000
* maintain/require county borders
* aggregate by closest geographic centroid
* do not exclude any areas
}
}
\keyword{datasets}
