% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyGATaggregators.R
\name{identifyGATaggregators}
\alias{identifyGATaggregators}
\title{Identify and Check GAT Aggregation Variables}
\usage{
identifyGATaggregators(
  shp,
  step = 4,
  agglist = NULL,
  backopt = TRUE,
  bgcol = "lightskyblue3",
  quitopt = "Quit",
  buttoncol = "cornflowerblue"
)
}
\arguments{
\item{shp}{Spatial layer.}

\item{step}{Integer step in GAT, for help reference.}

\item{agglist}{List of aggregation variables, with minimum and maximum values,
if pre-selected.}

\item{backopt}{Boolean denoting whether to include the back button.}

\item{bgcol}{Text string containing UI background color.}

\item{quitopt}{Text string for the cancel button.}

\item{buttoncol}{Text string containing UI button color.}
}
\description{
This function opens a dialog window for the user to select which variables
in the shapefile should be aggregated and what their minimum and maximum
values should be using the function \code{\link{inputGATaggregators}}. It
checks the resulting numeric values to ensure they are numbers and if
"NONE" is selected for the second aggregation variable, it assigns the
values for the first aggregation variable to the values for the second
aggregation variable.
}
\examples{

if (interactive()) {
identifyGATaggregators(shp = hftown)
}

}
