% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGATweights.R
\name{importGATweights}
\alias{importGATweights}
\title{Create Population Base for Weighting}
\usage{
importGATweights(area, filein, pathin, popvar = "Pop_tot")
}
\arguments{
\item{area}{The spatial polygons data frame you want to aggregate.}

\item{filein}{The filename of the population shapefile.}

\item{pathin}{The path of the population shapefile.}

\item{popvar}{The population variable to use in the population
shapefile.}
}
\description{
This function reads in the layer to be aggregated (large area) and the
file details for the shapefile containing small areas, including the file
name, file location, and variable containing population values.

It subsets the small area layer based on boundaries of the large are layer,
then calculates the proportion of each small area that falls in each large
area based on the area proportion of the small area that falls in each
large area. This creates a base layer from which to create population
weighted centroids.
}
\details{
This function has not yet been incorprated into the default version of
GAT, but I am including it here for testing. This help file and the
default settings for variables will need to be fleshed out more,
with examples provided, later.

I also need to decide if I will include a census block file in the
package, but I would prefer not to, mostly to keep filesize down.
}
\examples{

\donttest{
pathin <- paste0(find.package("gatpkg"), "/extdata")
mywtshp <- importGATweights(shp = hftown, filein = "hfblock", pathin)
}

}
