% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGATcompare.R
\name{plotGATcompare}
\alias{plotGATcompare}
\title{Plot GAT Comparison Map}
\usage{
plotGATcompare(areaold, areanew, mergevars, gatvars)
}
\arguments{
\item{areaold}{A spatial polygons data frame.}

\item{areanew}{A second spatial polygons data frame that ahould have the
same outer boundary as the first one.}

\item{mergevars}{A list of settings for the aggregation, including type of
aggregation (mergeopt1) and, if relevant, the variables
to compare, similar1 and similar2.}

\item{gatvars}{A list of objects created by the GAT tool. It contains the
strings aggregator1 and aggregator2, which are numeric
variables in the area, and the numbers minvalue1 and
minvalue2.}
}
\description{
This function draws two maps on top of each other with only the polygon
boundaries visible, so that polygon sizes and compositions can be compared.
It is designed to be saved to a recordPlot object for later rendering in a
PDF rather than displayed directly.
}
\details{
If plotting directly, the display may be distorted. For best results, save
to a recordPlot object and write to PDF.
}
\examples{
# define mapping variables
gatvars <- list(
  myidvar = "ID",             # character variable of unique values
  aggregator1 = "TOTAL_POP",  # numeric variable
  aggregator2 = "TOTAL_POP",  # numeric variable
  minvalue1 = 5000, minvalue2 = 5000,
  maxvalue1 = 15000, maxvalue2 = 15000,
  boundary = "COUNTY"         # character variable of non-unique values
)

# define merge type
mergevars <- list(
  mergeopt1 = "similar",    # can be similar, closest, or least
  similar1 = "B_TOT",       # numeric variable
  similar2 = "W_TOT",       # numeric variable without any zeros
  centroid = "geographic"
)

# draw the map
plotGATcompare(
  areaold = hftown,
  areanew = hfagg5k,
  mergevars = mergevars,
  gatvars = gatvars
)

}
