% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGATmaps.R
\name{plotGATmaps}
\alias{plotGATmaps}
\title{Plot GAT Maps}
\usage{
plotGATmaps(
  area,
  var,
  clr = "Blues",
  title.main = "",
  class = NULL,
  after = FALSE,
  title.sub = NULL,
  breaks = 5,
  colcode = NULL,
  mapstats = FALSE,
  ratemap = FALSE,
  closemap = FALSE
)
}
\arguments{
\item{area}{A spatial polygons data frame.}

\item{var}{A numeric variable within the spatial polygons data frame.}

\item{clr}{The color scheme to use for the map, from RColorBrewer.}

\item{title.main}{The map's main title as a text string.}

\item{class}{The cut points for the choropleth categories. Default is
NULL, in which case the function calculates breaks from
the data using the breaks value.}

\item{after}{Boolean denoting whether the function is mapping
the aggregated area "after" map.}

\item{title.sub}{The map's subtitle (if relevant) as a text string.}

\item{breaks}{Number of categories for the choropleth.}

\item{colcode}{Color code predefined for before/after maps.}

\item{mapstats}{A boolean denoting whether or not to include summary
statistics: min, median, and max.}

\item{ratemap}{Boolean denoting whether the variable that is to
be plotted contains decimal values that require rounding.
In GAT, this applies to the GAT-calculated rate.}

\item{closemap}{A boolean to denote whether to close the map window after
the map is drawn and saved.}
}
\description{
This function draws a choropleth map of a numeric variable. It includes
options to display summary statistics and subtitle. The legend produced is
dependent on the dataset; if there are no exclusions, no areas below the
minimum aggregation value, and no areas with missing rates, their respective
entries will be excluded from the legend.
}
\details{
The map produced by this function is designed to be saved to a recordPlot
object for later rendering in a PDF rather than displayed directly.
If plotting directly, the display may be distorted. For best results, save
to a recordPlot object and write to PDF.
}
\examples{

## plot population
# create a randomized flag variable (will not match label)
hftown@data$GATflag <-
  sample(
    c(0, 10), nrow(hftown@data),
    replace = TRUE,
    prob = c(.9, .1)
  )
# plot the variable
plotGATmaps(
  area = hftown,
  var = "TOTAL_POP",
  clr = "YlGn",
  title.main = "Population",
  breaks = 3,
  mapstats = TRUE
)

## plot a rate
# create a randomized flag variable (will not match labels)
# 0 = no flag, 1-3 = excluded by user, 10 = below minimum aggregation value
hftown@data$GATflag <-
  sample(
    c(0, 1, 10), nrow(hftown@data),
    replace = TRUE,
    prob = c(.8, .1, .1)
  )

# create a rate
hftown@data$pop_dens <- 100 * hftown@data$TOTAL_POP / hftown@data$AREALAND
# add some random missings to illustrate uncalculated rates
hftown@data$pop_dens[sample(length(hftown@data), 2)] <- NA

# plot the rate
plotGATmaps(
  area = hftown,
  var = "pop_dens",
  clr = "YlGn",
  title.main = "population density",
  breaks = 3,
  mapstats = TRUE,
  ratemap = TRUE
)

}
