% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGATprogram.R
\name{runGATprogram}
\alias{runGATprogram}
\title{Run the Geographic Aggregation Tool (GAT)}
\usage{
runGATprogram(
  limitdenom = FALSE,
  pwrepeat = FALSE,
  settings = NULL,
  adjacent = TRUE,
  minfirst = FALSE,
  closemap = FALSE
)
}
\arguments{
\item{limitdenom}{A boolean denoting whether to force denominators in rates
and merge ratios to contain only non-zero values.}

\item{pwrepeat}{A boolean denoting whether population weighting (if used)
should be recalculated each time two areas are merged
(TRUE) or if area centroids should be weighted with area
populations (FALSE). If population weighting is not used,
this option is ignored. If population weighting is used,
a polygon shapefile containing population is needed.}

\item{settings}{The filepath to an Rdata file previously created by GAT.}

\item{adjacent}{A boolean denoting whether to force GAT to merge only
adjacent areas.}

\item{minfirst}{A boolean denoting whether or not to select the most
desirable neighbor only from among the neighbors that
have values below the desired minimum. If no neighbors
are below the desired minimum, the most desirable of all
eligible neighbors is selected.}

\item{closemap}{A boolean to denote whether to close the map windows after
the maps are drawn and saved. The default setting is TRUE
and will result in maps remaining open after GAT is
finished running. Maps are saved to PDF irrespective of
this setting.}
}
\description{
This function runs GAT and does not require any inputs. The pop-up dialogs
it creates guide the user to define all parameters. For step by step details
on how GAT works, browse the package vignettes, especially the GAT manual,
which you can access via
\href{../doc/gat_tutorial.html}{
\code{vignette("gat_tutorial", package = "gatpkg")}}.
}
\details{
For more information  on how the different merge options work, see
"How GAT identifies areas to merge" in the
\href{../doc/gat_tech_notes.html}{
\code{vignette("gat_tech_notes", package = "gatpkg")}}.
}
\examples{

if (interactive()) {
# this code will run the default version of GAT
runGATprogram()
}

}
