% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightGATmap.R
\name{weightGATmap}
\alias{weightGATmap}
\title{Find the Weighted Centriods for the Map}
\usage{
weightGATmap(area, pop, idvar, popvar, crs = NULL)
}
\arguments{
\item{area}{A spatial layer representing areas to be aggregated.}

\item{pop}{A spatial layer containing underlying population values.}

\item{idvar}{A variable of unique string values to identify the area
layer's observations.}

\item{popvar}{The base population variable.}

\item{crs}{A user-defined non-lat/long projection, entered as a
string. The default is NULL, in which case the function
defines the projection.}
}
\description{
This function identifies the population weighted centroids for all areas
in your original map.
}
\examples{

if (interactive()) {
mycentroids <-
  weightGATmap(
    area = hftown, pop = hfpop,
    idvar = "ID", popvar = "Pop_tot")
}


}
