% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightGATregion.R
\name{weightGATregion}
\alias{weightGATregion}
\title{Find the Weighted Centriod for a GAT Region}
\usage{
weightGATregion(area, pop, IDlist, idvar, nrid)
}
\arguments{
\item{area}{The original shapefile.}

\item{pop}{The base population shapefile.}

\item{IDlist}{The list of all GAT IDs and their corresponding area IDs.}

\item{idvar}{A variable of unique string values to identify the
observations.}

\item{nrid}{A string that will be the ID for the merged observation
(polygon).}
}
\description{
This function identifies the population weighted centroid of a new
area based on the GAT ID assigned to the old areas and the population
shapefile.
}
\examples{

\donttest{
ids <- c("e", "e", "d", "d", "f", "f", "f", "e", "e", "d", "e",
         "d", "b", "c", "c", "a", "b", "c", "a", "b", "c")
mycentroid <-
  weightGATregion(
    area = hftown,
    pop = hfpop,
    IDlist = ids,
    idvar = "ID",
    nrid = "c"
  )
}


}
