% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatpkg.R
\docType{package}
\name{gatpkg}
\alias{gatpkg}
\alias{_PACKAGE}
\alias{GAT}
\alias{GAT's}
\alias{goals:}
\alias{\itemize{}
\alias{\item}
\alias{simplify}
\alias{geographic}
\alias{aggregation}
\alias{be}
\alias{in}
\alias{a}
\alias{format}
\alias{accessible}
\alias{to}
\alias{anyone}
\alias{not}
\alias{require}
\alias{prior}
\alias{experience}
\alias{with}
\alias{R}
\alias{customizable}
\alias{for}
\alias{advanced}
\alias{users}
\alias{}}
\alias{To}
\alias{run}
\alias{the}
\alias{default}
\alias{version}
\alias{of}
\alias{GAT,}
\alias{follow}
\alias{user's}
\alias{manual}
\alias{vignettes:}
\alias{`placeholder}
\alias{-}
\alias{check}
\alias{link`.}
\alias{For}
\alias{further}
\alias{information}
\alias{about}
\alias{...}
\alias{`placeholder`}
\title{The NYS Geographic Aggregation Tool}
\description{
A method for easily aggregating spatial data using user-defined
parameters with minimal knowledge of or experience using R.
}
\author{
\bold{Authors} \cr
Abigail Stamm, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Gwen Babcock, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr

\bold{Alpha-testers} \cr
Maxime Fokoua Dongmo, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Arjita Rai, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Neil Muscatiello, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Marine Nalbandyan, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Catherine Wargo, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr

\bold{Other contributors} \cr
Douglas Done, Bureau of Environmental and Occupational Epidemiology,
New York State Department of Health \cr
Thomas Talbot, \cr
}
