% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyGATaggregators.R
\name{identifyGATaggregators}
\alias{identifyGATaggregators}
\title{Identify and Check GAT Aggregation Variables}
\usage{
identifyGATaggregators(mapdata, step = 4, agglist = NULL, backopt = TRUE)
}
\arguments{
\item{mapdata}{A data frame.}

\item{step}{An integer step in the GAT program, for help reference.}

\item{agglist}{The list of aggregation variables, with minimum and maximum
values, if pre-selected.}

\item{backopt}{Boolean denoting whether to include the back button.}
}
\description{
This function opens a dialog window for the user to select which variables
in the shapefile should be aggregated and what their minimum and maximum
values should be using the function \code{\link{inputGATaggregators}}. It
checks the resulting numeric values to ensure they are numbers and if
"NONE" is selected for the second aggregation variable, it assigns the
values for the first aggregation variable to the values for the second
aggregation variable.
}
\examples{

\donttest{
# identify variable to aggregate
identifyGATaggregators(
  mapdata = hftown@data
)
}

}
