% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGATprogram.R
\name{runGATprogram}
\alias{runGATprogram}
\title{Run GAT Program with one dialog for aggregators}
\usage{
runGATprogram(
  limitdenom = FALSE,
  pwrepeat = FALSE,
  settings = NULL,
  adjacent = TRUE,
  minfirst = FALSE
)
}
\arguments{
\item{limitdenom}{Boolean denoting whether to force denominators in rates
and merge ratios to contain only non-zero values.}

\item{pwrepeat}{A boolean denoting whether population weighting (if used)
should be recalculated each time two areas are merged
(TRUE) or if area centroids should be weighted with area
populations (FALSE). If population weighting is not used,
this option is ignored. If population weighting is used,
a polygon shapefile containing population is needed.}

\item{settings}{The filepath to an Rdata file previously created by GAT.}

\item{adjacent}{A boolean denoting whether to force GAT to merge only
adjacent areas.}

\item{minfirst}{A boolean denoting whether or not to select the most
desirable neighbor only from among the neighbors that
have values below the desired minimum. If no neighbors
are below the desired minimum, the most desirable of all
elligible neighbors is selected.}
}
\description{
This function handles nearly the entire program, except for installing and
loading packages. It does not require any inputs, as the popup dialogs it
creates guide the user to define all parameters. For step by step details
on how this program works, browse the package vignettes, especially the
GAT manual, which you can access via
\href{../doc/gat_tutorial.html}{
\code{vignette("gat_tutorial", package = "gatpkg")}}.
}
\details{
You can load "BWidget" in advance, which opens a more navigable shapefile
selection window, but it is not necessary to run GAT.
}
\examples{

\donttest{
# this code will run the default version of GAT
runGATprogram() # uses defaults
}

}
