% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGAThelp.R
\name{showGAThelp}
\alias{showGAThelp}
\title{Show GAT Help Pages}
\usage{
showGAThelp(
  help = "Find help here.",
  helppage = "showGAThelp",
  step = 0,
  helptitle = "this step",
  helpimg = "showGAThelp"
)
}
\arguments{
\item{help}{A text string containing the help message.}

\item{helppage}{A text string that contains the funcion name for the
relevant function (if any) in the help dialog.}

\item{step}{Integer step in the GAT program, for help reference.}

\item{helptitle}{The step name to display in the title bar.}

\item{helpimg}{A text string denoting the file name of the image to be
shown.}
}
\description{
This function creates a dialog box to provide the user with additional
assistance in using GAT, including instructions and links to help pages.
}
\examples{

if (interactive()) {
# create the help message
hlp <- paste0("Instructions: \n",
              "  \u2022  To continue,  click 'Next >'. \n",
              "  \u2022  To return to the last option, click '< Back'. \n",
              "  \u2022  To quit GAT, click 'Cancel'.")

# create the dialog box
showGAThelp(help = hlp, helppage = "showGAThelp", step = 0,
            helptitle = "GAT help dialog")
}

}
