% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightGATmap.R
\name{weightGATmap}
\alias{weightGATmap}
\title{Find the Weighted Centriods for the Map}
\usage{
weightGATmap(area, pop, idvar, popvar, crs = NULL)
}
\arguments{
\item{area}{Spatial layer representing areas to be aggregated.}

\item{pop}{Spatial layer containing underlying population values.}

\item{idvar}{Variable of unique string values to identify the layer's
observations.}

\item{popvar}{The base population variable.}

\item{crs}{User-defined non-lat/long projection, entered as a string.
For the default, NULL, the function defines the projection.}
}
\description{
This function identifies the population weighted centroids for all areas
in your original map.
}
\examples{

if (interactive()) {
cen <- weightGATmap(area = hftown, pop = hfpop, idvar = "ID", popvar = "Pop")
}

}
