% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightGATmap.R
\name{weightGATmap}
\alias{weightGATmap}
\title{Find the Weighted Centriods for the Map}
\usage{
weightGATmap(area, filevars, idvar, popvar)
}
\arguments{
\item{area}{The original shapefile.}

\item{filevars}{A list of strings denoting file names and paths.}

\item{idvar}{A variable of unique string values to identify the
observations.}

\item{popvar}{The base population variable.}
}
\description{
This function identifies the population weighted centroids for all areas
in your original map.
}
\examples{

\donttest{
filevars <- list(
  poppath = paste0(find.package("gatpkg"), "/extdata"),
  popfile = "hfblock"
)

mycentroids <-
  weightGATmap(
    area = hftown,
    filevars = filevars,
    idvar = "ID",
    popvar = "Pop_tot"
  )
}


}
